threads :index , store , update , destroy with tests

+++ app/Http/Controllers/API/v1/Channels/ChannelController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\API\v1\Channels;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Channel\ChannelRequest;
 use App\Http\Requests\Channel\DeleteChannelRequest;
+use App\Models\Channel;
 use App\Models\Permission;
 use App\Repositories\ChannelRepo;
+use App\Responses\AjaxResponse;
 use Illuminate\Http\Response;
 
 class ChannelController extends Controller
@@ -16,35 +18,33 @@ class ChannelController extends Controller
     public function __construct(ChannelRepo $channelRepo)
     {
         $this->channelRepo = $channelRepo;
-
-        $this->middleware(['can:' . Permission::PERMISSION_MANAGE_CHANNELS])->only('store');
-        $this->middleware(['can:' . Permission::PERMISSION_MANAGE_CHANNELS])->only('update');
-        $this->middleware(['can:' . Permission::PERMISSION_MANAGE_CHANNELS])->only('destroy');
     }
 
     public function getAllChannelList()
     {
         $channels = $this->channelRepo->all();
-        return response()->json($channels, Response::HTTP_OK);
+        return AjaxResponse::SendData($channels);
     }
 
     public function store(ChannelRequest $request)
     {
+        $this->authorize('store' , Channel::class);
         $this->channelRepo->store($request);
-        return response()->json(['message' => 'channel created successfully'], Response::HTTP_CREATED);
+        return AjaxResponse::created('channel created successfully');
     }
 
     public function update(ChannelRequest $request)
     {
+        $this->authorize('update' , Channel::class);
         $this->channelRepo->update($request);
-        return response()->json(['message' => 'channel updated successfully'], Response::HTTP_OK);
+        return AjaxResponse::ok('channel updated successfully');
     }
 
     public function destroy(DeleteChannelRequest $request)
     {
+        $this->authorize('delete' , Channel::class);
         $channel = $this->channelRepo->findById($request->id);
         $channel->delete();
-        return \response(['message' => 'channel deleted successfully'], Response::HTTP_OK);
-
+        return AjaxResponse::ok('channel deleted successfully');
     }
 }
